--- apps/augury/src/atoms/onboarding/atom.tsx
@@ -1,11 +1,0 @@
-import { atom } from 'recoil';
-
-export const OnboardingPage = atom({
-  key: 'onboardingPage',
-  default: 0,
-})
-
-export const CustomComposition = atom({
-  key: 'onboardingCustomComposition',
-  default: false,
-})
