+++ apps/augury/src/components/navigation/NavbarIcon.tsx
@@ -5,33 +5,33 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 interface Props {
   icon: IconDefinition;
   iconActivated?: IconDefinition;
-  text?: string;
-  activated?: boolean;
-  link: string; // Unsure how to manage this
+  text: string;
+  activated: boolean;
+  link: string;
 }
 
 export default function PageIcon(props: Props): JSX.Element {
   return (
     <Link
       flex="1"
-      bg={props?.activated ? 'background.mantle' : 'background.crust'}
-      color={props?.activated ? 'text.subtitle0' : 'text.subtitle1'}
+      bg={props.activated ? 'background.mantle' : 'background.crust'}
+      color={props.activated ? 'text.subtitle0' : 'text.subtitle1'}
       _hover={{
         bg: 'background.selBg',
         color: 'text.body',
       }}
-      href={props?.link}
+      href={props.link}
     >
-      <VisuallyHidden>{props?.text}</VisuallyHidden>
+      <VisuallyHidden>{props.text}</VisuallyHidden>
       <Center h="full">
         <FontAwesomeIcon
           icon={
-            props?.activated && props?.iconActivated
-              ? props?.iconActivated
-              : props?.icon
+            props.activated && props.iconActivated
+              ? props.iconActivated
+              : props.icon
           }
           size="xl"
-        ></FontAwesomeIcon>
+        />
       </Center>
     </Link>
   );
