+++ apps/augury/src/components/navigation/navbarPagesData.ts
@@ -1,0 +1,23 @@
+import { faHouse, faBriefcase, faGear } from "@fortawesome/free-solid-svg-icons";
+import { NavbarPage } from "./Navbar";
+
+export const navbarPages: NavbarPage[] = [
+  {
+    icon: faHouse,
+    link: '/',
+    text: 'home',
+    baseValue: 0,
+  },
+  {
+    icon: faBriefcase,
+    link: '/portfolio',
+    text: 'portfolio',
+    baseValue: 1,
+  },
+  {
+    icon: faGear,
+    link: '/settings',
+    text: 'settings',
+    baseValue: 2,
+  },
+];
